work on role auth

+++ src/services/posts.service.js
@@ -1,5 +1,5 @@
-const { NotFoundError } = require('../models/errors');
-const ObjectId = require('mongodb').ObjectId;
+const { NotFoundError, AuthorizationError } = require('../models/errors');
+const { ObjectId } = require('mongodb');
 const db = require('../mongo').db();
 
 let posts = db.collection('posts');
@@ -20,8 +20,7 @@ const createPost = async (post) => {
 }
 
 const getPosts = async () => {
-    let posts = await posts.find({}).toArray();
-    return posts;
+    return await posts.find({}).toArray();
 }
 
 const getUserPosts = async (user_id) => {
@@ -31,11 +30,15 @@ const getUserPosts = async (user_id) => {
     return user_posts;
 }
 
-const delPost = async (post) => {
+const delPost = async (post, current_user) => {
     const query = { _id: ObjectId(post._id)};
 
-    let existing_post = await users.findOne(query);
+    let existing_post = await posts.findOne(query);
     if (!existing_post) throw new NotFoundError('Post does not exist');
+
+    if (!existing_post.owner.equals(current_user._id) && !current_user.roles.includes('admin')) {
+        throw new AuthorizationError("Action not permitted");
+    }
 
     await posts.deleteOne(query);
     return;
